Chessboard
Write a program that creates a string that represents an 8x8 grid, using new-line characters to separate lines. At each position of the grid there is either a space of a # character. The characters should form a chessboard (see above expected output).

When you have a program that generates this pattern, define a binding size = 8 and change the program so that it works for any size, outputting a grid of the given width and height.